refactor(jsdoc): getSwitches
