+++ @types/window.d.ts
@@ -1,5 +1,5 @@
 import type { TeamSubscription } from '../libs/@guardian/libs/src/logger/log.js';
-import type { Switches } from '../libs/@guardian/libs/src/switches/@types/Switches';
+import type { Switches } from '../libs/@guardian/libs/src/switches/getSwitches.js';
 
 declare global {
 	interface Window {
