--- libs/@guardian/libs/src/switches/@types/Switches.ts
@@ -1,1 +1,0 @@
-export type Switches = Record<string, boolean>;
